requirements.txt
rshell>=0.0.31
esptool>=4.8.dev1
micropython-esp32-stubs>=1.20.0.*
